+++ libavcodec/h263dsp.c
@@ -28,7 +28,7 @@ const uint8_t ff_h263_loop_filter_strength[32] = {
     7, 8, 8, 8, 9, 9, 9, 10, 10, 10, 11, 11, 11, 12, 12, 12
 };
 
-static void h263_h_loop_filter_c(uint8_t *src, int stride, int qscale)
+static void h263_h_loop_filter_c(uint8_t *src, ptrdiff_t stride, int qscale)
 {
     int y;
     const int strength = ff_h263_loop_filter_strength[qscale];
@@ -71,7 +71,7 @@ static void h263_h_loop_filter_c(uint8_t *src, int stride, int qscale)
     }
 }
 
-static void h263_v_loop_filter_c(uint8_t *src, int stride, int qscale)
+static void h263_v_loop_filter_c(uint8_t *src, ptrdiff_t stride, int qscale)
 {
     int x;
     const int strength = ff_h263_loop_filter_strength[qscale];
