+++ libavcodec/mips/h263dsp_mips.h
@@ -23,8 +23,8 @@
 
 #include "libavcodec/mpegvideo.h"
 
-void ff_h263_h_loop_filter_msa(uint8_t *src, int stride, int q_scale);
-void ff_h263_v_loop_filter_msa(uint8_t *src, int stride, int q_scale);
+void ff_h263_h_loop_filter_msa(uint8_t *src, ptrdiff_t stride, int q_scale);
+void ff_h263_v_loop_filter_msa(uint8_t *src, ptrdiff_t stride, int q_scale);
 void ff_dct_unquantize_mpeg2_inter_msa(const MPVContext *s, int16_t *block,
                                        int32_t index, int32_t q_scale);
 void ff_dct_unquantize_h263_inter_msa(const MPVContext *s, int16_t *block,
